requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.16
pandas==1.5.3
openpyxl==3.1.2
pytz==2023.3
python-dateutil==2.8.2
gunicorn==21.2.0
Flask-WTF
schedule
concurrent-log-handler
redis

# --- Database ---
psycopg2-binary==2.9.9  # Dùng binary cho dễ cài đặt trên linux/windows
pandas==1.5.3           # Nếu bạn cần xử lý dữ liệu excel
openpyxl==3.1.2

# --- Security & Utils ---
email-validator==2.0.0
python-dotenv==1.0.0    # Đọc file .env
pytz==2023.3
python-dateutil==2.8.2

# --- 2FA & Notification ---
Flask-Mail==0.9.1       # Để gửi OTP qua Email
twilio==8.5.0           # Để gửi OTP qua SMS (tùy chọn)

# --- Deployment ---
gunicorn==21.2.0
schedule==1.2.0